fix(webui): Provider 保存与连通测试 API 校验修复
PUT providers 忽略 GET 只读元数据；latency_ms 接受 AI 仓返回的小数，避免测试 500。

+++ packages/pb_webui/console_openapi_models.py
@@ -131,7 +131,7 @@ class LlmProviderTestData(BaseModel):
     ok: bool
     provider_id: str = ""
     model: str = ""
-    latency_ms: int | None = None
+    latency_ms: float | None = None
     error: str = ""
     detail: str = ""
